requirements.txt
streamlit
deepface
opencv-python
tensorflow
numpy
Pillow
ultralytics
omegaconf
tf-keras
tensorflow
opencv-python-headless
dlib